+++ Runtime/Scripts/Environment.cs
@@ -1,0 +1,12 @@
+﻿namespace KindMen.Uxios
+{
+    public interface IEnvironment
+    {
+        
+    }
+
+    public class Environment<TPayload> : IEnvironment where TPayload : class
+    {
+        public TPayload Payload { get; set; }
+    }
+}