+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.3] - 2026-01-12
+
+### Added
+
+- Environment interface and payload that can be added to the config. This will allow extra information to be associated with
+  a request, reducing the need for capturing in promise lambdas.
+
 ## [0.4.2] - 2025-12-03
 
 ### Fixed